requirements.txt
appscript==1.3.0
asgiref==3.8.1
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.4.26
charset-normalizer==3.4.2
cobble==0.1.4
Django==5.2.2
docx2pdf==0.1.8
idna==3.10
lxml==5.4.0
mammoth==1.9.1
numpy==2.2.6
opencv-python==4.11.0.86
packaging==25.0
pdf2image==1.17.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyMuPDF==1.26.0
pytesseract==0.3.13
python-docx==1.1.2
python-dotenv==1.1.0
python-pptx==1.0.2
requests==2.32.3
soupsieve==2.7
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
XlsxWriter==3.2.3
